Create recipe for jenkins1.o.o that replicates roles
This moves all of the logic we have currently in roles for jenkins1.osuosl.org
into it's own recipe to test it easier. This also moves specific changes for
jenkins1 out of the master recipe. The master recipe was pulling in things it
shouldn't be such as chefdk due to the inclusion of the cookbook uploader
recipe. This simplifies it for easier testing of other new recipes down the
road.

In addition, I also made these changes and/or fixes:

- Updated the repo to use to test to test-cookbook (lanparty was removed)
- Created a github security group on OpenStack so that we don't leave these
  instance wide open to the internet
- Added very basic unit tests for the jenkins1 recipe
- Created a shared ServerSpec test for testing basic jenkins functionality
- Include build-essential for CentOS 6 due to an upstream issue with the jenkins
  cookbook