Upgrade to jenkins-2.32.3
This is a major upgrade from 1.x to 2.x but shouldn't cause too many issues on
current systems as long as you run any updates/fixes it needs. This is dependant
on osuosl-cookbooks/osl-munin#19 in order to use the jenkins-5.0.0 cookbook. The
newer cookbook is required to add support for Jenkins 2.x.

New fixes:

- Added a ChefSpec test to ensure the right version of jenkins is being
  installed
- Added two ServerSpec curl tests to ensure http redirects properly and jenkins
  is running and showing the correct version
- Added the missing spec_helper require on the master suite for ServerSpec
- Fixes for ssh credentials in Jenkins 2.x
- Switch to using jenkins_command[safe-restart] instead of service to make the
  restart work more smoothly
- Improve ChefSpec tests to test for notifications