Sanely restart jenkins before you use plugins later in the run
This provides a trigger mechanism based on touching a file when a plugin gets
installed or updated. After all of the plugins have been installed, the log
resource will trigger if it sees the file and immediately restart jenkins so
that the plugins load up.

This should clean up some annoying issues we've had in the past on powerci.