Always upgrade openstack_taster
Since we control when new releases of openstack_taster, we don't need to lock
the version in the cookbook. This removes the requirement of bumping this
cookbook on every new release. It also ensures that it always uses the latest
version which it currently does even if you bump the version.