Set anonymous with admin privileges in a test cookbook
This is ONLY needed for test-kitchen and should NOT be used for production for
obvious reasons. This is needed in cases where we use external auth (such as
github) which requires you to use your auth. This allows ServerSpec tests to
work a little better an saner in those cases.