Set alfred gid/uid to something static
This is needed for utilizing the docker-custom-build-environment plugin. We need
to have a known static UID in containers on all hosts acting as a jenkins node.
I will be deploying a one-time manual fix on each current node as this doesn't
fix any current running processes or change the gid for the home directory
correctly from the current setup.