Convert to using users_manage resource
This replaces using user, group resources and the ssh-keys cookbook. This change
is needed due to issues using newer ChefSpec with the ssh-keys cookbook. Also,
it's been in our long term plan to migrate away from the user cookbook to the
users cookbook.

This change will require a similar change in the production data bag for the
alfred user.