Use vhost hostname for jenkins server webhooks
By default, we're using the FQDN for setting the hostname used in Github
webhooks. This becomes a problem when deploying more than one OSL Jenkins server
for a migration. It makes more sense to just point to the hostname we actually
use for our Jenkins instance.

This should make it easier to do major server platform upgrades on our primary
Jenkins server.