Allow for manual upgrades of Jenkins and plugins
This is a temporary fix which disables all use of jenkins resources and locking
of the jenkins version to install. This will allow us to at least get the
current systems updated.

I commented out any code we don't want running currently and will be fixed in a
later PR properly.

Signed-off-by: Lance Albertson <lance@osuosl.org>