Major refactor
This is a major refactor of how to deploy Jenkins servers and it's related
services. This moves us away from the upstream jenkins cookbook and utilize our
own resources.

This refactor uses Jenkins Configuration as Code [1] along with the Plugin
Installation Manager Tool for Jenkins [2] for configuring various aspects.

In addition, there's some other changes:

- Removed recipes and code related to the packer pipeline
- Remove various recipes that have been converted to resources
- Remove site PR builder recipe as we don't really use it like this anymore
- Don't deploy test-kitchen instances on the public network by default. This is
  only needed on powerci and ibmz-ci tests due to Github OAuth
- Use Haproxy resources directly so we can correctly setup the forward proxy
- Remove use of the yum versionlock plugin
- Migrate to proper InSpec test suites

[1] https://plugins.jenkins.io/configuration-as-code/
[2] https://github.com/jenkinsci/plugin-installation-manager-tool

Signed-off-by: Lance Albertson <lance@osuosl.org>